crawler_AI/requirements.txt
# Requirements for Domain Crawler
# Install with: pip install -r requirements.txt

# Core dependencies
playwright>=1.40.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
aiofiles>=23.0.0
dnspython>=2.4.0
requests>=2.31.0
tldextract>=5.0.0
asyncio-throttle>=1.0.2

# Optional but recommended external tools
# Install separately:
# - subfinder: go install github.com/projectdiscovery/subfinder/v2/cmd/subfinder@latest
# - amass: go install github.com/owasp-amass/amass/v4/...@master

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Additional utilities
colorlog>=6.7.0
tqdm>=4.65.0
pyyaml>=6.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# After installation, run:
# playwright install chromium